src/requirements.txt
flask
pillow
requests
PyYAML